requirements.txt
datasets==2.18.0
Flask==3.0.2
pandas==2.2.1
peft==0.7.1
trl==0.7.4
torch==2.2.0+cu118
torchaudio==2.2.0+cu118
torchdata==0.7.1
torchtext==0.17.0
torchvision==0.17.0+cu118
transformer==4.36.2